refactor(io): move internal/secfile to internal/core/io/secfile (#210)
