+++ internal/core/io/secfile/secfile.go
