+++ internal/core/io/secfile/secfile_test.go
